requirements.txt
Django~=3.2
django-cors-headers~=3.4.0
djangorestframework~=3.11.1
djangorestframework-simplejwt==4.5.0
PyJWT==1.7.1
dj-rest-auth==2.1.3
django-allauth==0.44.0
